Refactor enhancedImageProcessor.js into modular structure
This commit refactors the monolithic enhancedImageProcessor.js into multiple modules:
- ImageProcessor: Handles image analysis and processing
- MetadataExtractor: Extracts patent details from HTML files
- FileSystemHelper: Manages file operations
- config.js: Centralizes configuration values

The original enhancedImageProcessor.js now acts as a wrapper
that imports and exports from the modular implementation,
ensuring backward compatibility with existing code.
